move summarize_circuits to circuit_utils
